requirements.txt
streamlit==1.33.0
streamlit-lottie==0.0.5
streamlit-folium==0.15.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
folium==0.16.0
plotly==5.20.0
